pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ez-comfy"
version = "0.1.0"
description = "Intelligent ComfyUI orchestrator — prompt to optimal workflow"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "websockets>=12.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "psutil>=5.9",
    "pillow>=10.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
llm = [
    "openai>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
]

[project.scripts]
ez-comfy = "ez_comfy.__main__:main"

[tool.setuptools.packages.find]
include = ["ez_comfy*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
